requirements.txt
python-dotenv
fastapi
openai>=0.27.0
starlette
pydantic
uvicorn
